+++ packages/juice_paging/lib/src/use_cases/initialize_paging_use_case.dart
@@ -9,7 +9,7 @@ class InitializePagingUseCase<T>
   @override
   Future<void> execute(InitializePagingEvent event) async {
     if (bloc.config.loadOnInit) {
-      bloc.send(RefreshPageEvent());
+      await bloc.send(RefreshPageEvent());
     }
   }
 }